+++ gatsby-browser.js
@@ -1,6 +1,6 @@
 import React from "react"
 
-import { DarkThemeProvider } from './src/pages/components/DarkThemeContext/DarkThemeContext'
+import { DarkThemeProvider } from './src/components/DarkThemeContext/DarkThemeContext'
 
 export const wrapRootElement = ({ element }) => {
   return <DarkThemeProvider>{element}</DarkThemeProvider>
